Moved review child components into their own separate folders
