+++ client/src/components/reviews/ReviewList/ReviewList.jsx
@@ -1,7 +1,7 @@
 import React, {useState, useEffect, useContext} from 'react';
 import axios from 'axios';
 import ReviewTile from './ReviewTile.jsx';
-import {FilterContext} from './Reviews.jsx';
+import {FilterContext} from '../Reviews.jsx';
 
 const ReviewList = (props) => {
   const [reviews, setReviews] = useState([]);
